getdataset, size_t

+++ FightingGameProject/ActiveControllers.cpp
@@ -24,7 +24,7 @@ namespace RB::Controllers
 
 	void ActiveControllers::UpdateAll()
 	{
-		for (int i = 0; i < _vecControllers.size(); i++)
+		for (size_t i = 0; i < _vecControllers.size(); i++)
 		{
 			_vecControllers[i]->OnUpdate();
 		}
@@ -32,7 +32,7 @@ namespace RB::Controllers
 
 	void ActiveControllers::FixedUpdateAll()
 	{
-		for (int i = 0; i < _vecControllers.size(); i++)
+		for (size_t i = 0; i < _vecControllers.size(); i++)
 		{
 			_vecControllers[i]->OnFixedUpdate();
 		}
@@ -40,7 +40,7 @@ namespace RB::Controllers
 
 	void ActiveControllers::_DestroyAll()
 	{
-		for (int i = 0; i < _vecControllers.size(); i++)
+		for (size_t i = 0; i < _vecControllers.size(); i++)
 		{
 			delete _vecControllers[i];
